Move utility files to utils submodule. Add camera setters. Update docs.
