+++ doc/uml/phosimClass.uml
@@ -3,12 +3,6 @@ package "telescope" #DDDDDD {
   class TeleFacade
 }
 
-package "ts_wep" #DDDDDD {
-  class WEPCalculation
-  class SourceProcessor
-  class WcsSol
-}
-
 package "ts_ofc" #DDDDDD {
   class OFCCalculation
 }
@@ -16,13 +10,9 @@ package "ts_ofc" #DDDDDD {
 PhosimCmpt ..> SkySim
 PhosimCmpt o-- OpdMetrology
 PhosimCmpt *-- TeleFacade
-SkySim o-- SourceProcessor
-SkySim o-- WcsSol
 CloseLoopTask o-- SkySim
-CloseLoopTask o-- WEPCalculation
 CloseLoopTask o-- OFCCalculation
 CloseLoopTask o-- PhosimCmpt
 CloseLoopTask ..> TeleFacade
 CloseLoopTask ..> OpdMetrology
-class CreatePhosimDonutTemplates
 @enduml
