+++ doc/uml/utilsClass.uml
@@ -1,0 +1,3 @@
+@startuml
+CreatePhosimDonutTemplates ..> MapSensorNameAndID
+@enduml
