+++ python/lsst/ts/phosim/utils/MapSensorNameAndId.py
